+++ WriteBalance.Domain/Entities/FinancialRecord.cs
@@ -15,8 +15,8 @@ namespace WriteBalance.Domain.Entities
         public string Moeen_Title { get; set; }
 
 
-        public long Mande_Bed { get; set; }
-        public long Mande_Bes { get; set; }
+        public decimal Mande_Bed { get; set; }
+        public decimal Mande_Bes { get; set; }
 
     }
 }
